control_plane/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx>=0.27.0
docker>=7.1.0
kafka-python>=2.0.2
neo4j>=5.18.0
qdrant-client>=1.9.0
hvac>=2.1.0
requests>=2.31.0
python-multipart>=0.0.12
websockets>=13.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
